Added primary display detection
Resolution should be selected based on the primary display as there might be multiple displays present.
This also solves the issue of the monitor interface being different for each system while it was hardcoded in the old commit for a specific use case that
would only work for a handful of users who use mainly laptops or other more integrated devices.

This commit adds a function to find the current primary display and uses it for getting the screen resolution to be used when rendering the wallpaper and when setting the rendered wallpaper to the screen

Fixes issue #1