Re-write text redner to fix resizing issue
This commit solves issue #2
Text would not resize currectly and if a large number of packages is present not all pacakges will fit in the image.

This is solved by rendering text using a web-browser capture insted of imagemagick.
This enables more flexability with the usage of javascript.
Imagmagic does not allow using both bestfit text size and coloring text at specific areas using pango.